# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEN Signal Intelligence API",
    "description": "## Signal Intelligence Engine\n\nOMEN transforms market events into structured probability signals.\n\n### Authentication\n\nAll API endpoints require authentication via API key.\nInclude the `X-API-Key` header in all requests.\n\n**Public endpoints (no auth required):**\n- `/health/*` - Health checks\n- `/docs`, `/redoc` - API documentation\n- `/metrics` - Prometheus metrics\n\n### What OMEN provides\n- Probability assessment with confidence bounds\n- Signal validation and quality scoring\n- Geographic and temporal context\n- Full evidence chain and traceability\n\n### What OMEN does NOT provide\n- Impact assessment (severity, delay, cost)\n- Decision steering (temporal proximity, relevance rank, high-confidence flags)\n- Recommendations or advice\n- Risk quantification\n\nImpact assessment is the responsibility of downstream systems.",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://omen.io/logo.png",
      "altText": "OMEN Logo"
    },
    "contact": {
      "name": "OMEN API Support",
      "url": "https://github.com/omen/omen",
      "email": "support@omen.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/partner-signals/": {
      "get": {
        "tags": [
          "Partner Signals",
          "Partner Signals"
        ],
        "summary": "Get signals for all monitored partners",
        "description": "Returns raw signal metrics for logistics partners.\n    \n    **IMPORTANT**: This endpoint returns SIGNALS only, not risk decisions.\n    \n    RiskCast should use these signals to make risk assessments based on:\n    - Order context\n    - User risk appetite\n    - Business rules\n    \n    **OMEN does NOT make risk decisions.**\n    \n    Response includes:\n    - `signals`: Raw metrics (price, volume, fundamentals)\n    - `evidence`: Evidence trail for audit\n    - `confidence`: Data quality indicators\n    - `omen_suggestion`: Optional suggestion (NOT a decision)\n    \n    Response does NOT include:\n    - `risk_status`\n    - `overall_risk`\n    - `risk_breakdown`",
        "operationId": "get_all_partner_signals_api_v1_partner_signals__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of symbols (default: GMD,HAH,VOS,VSC,PVT)",
              "title": "Symbols"
            },
            "description": "Comma-separated list of symbols (default: GMD,HAH,VOS,VSC,PVT)",
            "example": "HAH,GMD,VOS"
          },
          {
            "name": "include_evidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include signal evidence in response",
              "default": true,
              "title": "Include Evidence"
            },
            "description": "Include signal evidence in response"
          },
          {
            "name": "include_market_context",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include market context (VNINDEX, sector)",
              "default": false,
              "title": "Include Market Context"
            },
            "description": "Include market context (VNINDEX, sector)"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signals retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerSignalsListResponse"
                },
                "example": {
                  "timestamp": "2026-02-01T10:00:00Z",
                  "total_partners": 1,
                  "partners": [
                    {
                      "symbol": "GMD",
                      "company_name": "Gemadept Corporation",
                      "signals": {
                        "price_current": 68.5,
                        "price_previous_close": 67.9,
                        "price_change_percent": 0.88,
                        "volume": 1901300,
                        "volume_avg_20d": 1500000,
                        "volatility_20d": 0.023,
                        "liquidity_score": 0.85,
                        "trend_20d": 3.2,
                        "zscore_price": 0.45
                      },
                      "confidence": {
                        "overall_confidence": 0.91,
                        "data_completeness": 0.95,
                        "data_freshness_seconds": 120,
                        "price_data_confidence": 1.0,
                        "fundamental_data_confidence": 0.0,
                        "volume_data_confidence": 1.0,
                        "missing_fields": [],
                        "data_source": "vnstock",
                        "data_source_reliability": 0.95,
                        "confidence_interval": {
                          "point_estimate": 0.91,
                          "lower_bound": 0.85,
                          "upper_bound": 0.97,
                          "confidence_level": 0.95,
                          "method": "weighted_bayesian"
                        }
                      },
                      "evidence": [
                        {
                          "evidence_id": "GMD-PRICE-20260201100000",
                          "evidence_type": "PRICE_CHANGE",
                          "title": "GMD price increased 0.88%",
                          "raw_value": 0.88,
                          "normalized_value": 0.088,
                          "source": "vnstock",
                          "timestamp": "2026-02-01T10:00:00Z"
                        }
                      ],
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Data source unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-signals/partners": {
      "get": {
        "tags": [
          "Partner Signals",
          "Partner Signals"
        ],
        "summary": "List all monitored logistics partners",
        "description": "Returns metadata about each partner company.",
        "operationId": "list_partners_api_v1_partner_signals_partners_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Partners Api V1 Partner Signals Partners Get"
                },
                "example": {
                  "timestamp": "2026-02-01T10:00:00Z",
                  "total_partners": 1,
                  "partners": [
                    {
                      "symbol": "GMD",
                      "company_name": "Gemadept Corporation",
                      "signals": {
                        "price_current": 68.5,
                        "price_previous_close": 67.9,
                        "price_change_percent": 0.88,
                        "volume": 1901300,
                        "volume_avg_20d": 1500000,
                        "volatility_20d": 0.023,
                        "liquidity_score": 0.85,
                        "trend_20d": 3.2,
                        "zscore_price": 0.45
                      },
                      "confidence": {
                        "overall_confidence": 0.91,
                        "data_completeness": 0.95,
                        "data_freshness_seconds": 120,
                        "price_data_confidence": 1.0,
                        "fundamental_data_confidence": 0.0,
                        "volume_data_confidence": 1.0,
                        "missing_fields": [],
                        "data_source": "vnstock",
                        "data_source_reliability": 0.95,
                        "confidence_interval": {
                          "point_estimate": 0.91,
                          "lower_bound": 0.85,
                          "upper_bound": 0.97,
                          "confidence_level": 0.95,
                          "method": "weighted_bayesian"
                        }
                      },
                      "evidence": [
                        {
                          "evidence_id": "GMD-PRICE-20260201100000",
                          "evidence_type": "PRICE_CHANGE",
                          "title": "GMD price increased 0.88%",
                          "raw_value": 0.88,
                          "normalized_value": 0.088,
                          "source": "vnstock",
                          "timestamp": "2026-02-01T10:00:00Z"
                        }
                      ],
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-signals/{symbol}": {
      "get": {
        "tags": [
          "Partner Signals",
          "Partner Signals"
        ],
        "summary": "Get signals for a specific partner",
        "description": "Returns detailed signals for a single logistics partner.\n    \n    **IMPORTANT**: Returns SIGNALS only, not risk decisions.",
        "operationId": "get_partner_signal_api_v1_partner_signals__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "include_evidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include signal evidence",
              "default": true,
              "title": "Include Evidence"
            },
            "description": "Include signal evidence"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerSignalResponse"
                },
                "example": {
                  "symbol": "GMD",
                  "company_name": "Gemadept Corporation",
                  "signals": {
                    "price_current": 68.5,
                    "price_previous_close": 67.9,
                    "price_change_percent": 0.88,
                    "volume": 1901300,
                    "volume_avg_20d": 1500000,
                    "volatility_20d": 0.023,
                    "liquidity_score": 0.85,
                    "trend_20d": 3.2,
                    "zscore_price": 0.45
                  },
                  "confidence": {
                    "overall_confidence": 0.91,
                    "data_completeness": 0.95,
                    "data_freshness_seconds": 120,
                    "price_data_confidence": 1.0,
                    "fundamental_data_confidence": 0.0,
                    "volume_data_confidence": 1.0,
                    "missing_fields": [],
                    "data_source": "vnstock",
                    "data_source_reliability": 0.95,
                    "confidence_interval": {
                      "point_estimate": 0.91,
                      "lower_bound": 0.85,
                      "upper_bound": 0.97,
                      "confidence_level": 0.95,
                      "method": "weighted_bayesian"
                    }
                  },
                  "evidence": [
                    {
                      "evidence_id": "GMD-PRICE-20260201100000",
                      "evidence_type": "PRICE_CHANGE",
                      "title": "GMD price increased 0.88%",
                      "raw_value": 0.88,
                      "normalized_value": 0.088,
                      "source": "vnstock",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "timestamp": "2026-02-01T10:00:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-signals/{symbol}/price": {
      "get": {
        "tags": [
          "Partner Signals",
          "Partner Signals"
        ],
        "summary": "Get raw price data for a partner",
        "description": "Returns real-time price data without any risk assessment.",
        "operationId": "get_partner_price_api_v1_partner_signals__symbol__price_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Partner Price Api V1 Partner Signals  Symbol  Price Get"
                },
                "example": {
                  "symbol": "GMD",
                  "company_name": "Gemadept Corporation",
                  "signals": {
                    "price_current": 68.5,
                    "price_previous_close": 67.9,
                    "price_change_percent": 0.88,
                    "volume": 1901300,
                    "volume_avg_20d": 1500000,
                    "volatility_20d": 0.023,
                    "liquidity_score": 0.85,
                    "trend_20d": 3.2,
                    "zscore_price": 0.45
                  },
                  "confidence": {
                    "overall_confidence": 0.91,
                    "data_completeness": 0.95,
                    "data_freshness_seconds": 120,
                    "price_data_confidence": 1.0,
                    "fundamental_data_confidence": 0.0,
                    "volume_data_confidence": 1.0,
                    "missing_fields": [],
                    "data_source": "vnstock",
                    "data_source_reliability": 0.95,
                    "confidence_interval": {
                      "point_estimate": 0.91,
                      "lower_bound": 0.85,
                      "upper_bound": 0.97,
                      "confidence_level": 0.95,
                      "method": "weighted_bayesian"
                    }
                  },
                  "evidence": [
                    {
                      "evidence_id": "GMD-PRICE-20260201100000",
                      "evidence_type": "PRICE_CHANGE",
                      "title": "GMD price increased 0.88%",
                      "raw_value": 0.88,
                      "normalized_value": 0.088,
                      "source": "vnstock",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "timestamp": "2026-02-01T10:00:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-signals/{symbol}/fundamentals": {
      "get": {
        "tags": [
          "Partner Signals",
          "Partner Signals"
        ],
        "summary": "Get fundamental health indicators",
        "description": "Returns PE, ROE, and other fundamental metrics without risk classification.",
        "operationId": "get_partner_fundamentals_api_v1_partner_signals__symbol__fundamentals_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Partner Fundamentals Api V1 Partner Signals  Symbol  Fundamentals Get"
                },
                "example": {
                  "symbol": "GMD",
                  "company_name": "Gemadept Corporation",
                  "signals": {
                    "price_current": 68.5,
                    "price_previous_close": 67.9,
                    "price_change_percent": 0.88,
                    "volume": 1901300,
                    "volume_avg_20d": 1500000,
                    "volatility_20d": 0.023,
                    "liquidity_score": 0.85,
                    "trend_20d": 3.2,
                    "zscore_price": 0.45
                  },
                  "confidence": {
                    "overall_confidence": 0.91,
                    "data_completeness": 0.95,
                    "data_freshness_seconds": 120,
                    "price_data_confidence": 1.0,
                    "fundamental_data_confidence": 0.0,
                    "volume_data_confidence": 1.0,
                    "missing_fields": [],
                    "data_source": "vnstock",
                    "data_source_reliability": 0.95,
                    "confidence_interval": {
                      "point_estimate": 0.91,
                      "lower_bound": 0.85,
                      "upper_bound": 0.97,
                      "confidence_level": 0.95,
                      "method": "weighted_bayesian"
                    }
                  },
                  "evidence": [
                    {
                      "evidence_id": "GMD-PRICE-20260201100000",
                      "evidence_type": "PRICE_CHANGE",
                      "title": "GMD price increased 0.88%",
                      "raw_value": 0.88,
                      "normalized_value": 0.088,
                      "source": "vnstock",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "timestamp": "2026-02-01T10:00:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-risk/partners": {
      "get": {
        "tags": [
          "Partner Risk (DEPRECATED)",
          "Partner Risk (DEPRECATED)"
        ],
        "summary": "⚠️ DEPRECATED - Use /api/v1/partner-signals/partners",
        "description": "**⚠️ This endpoint is DEPRECATED.**\n    \n    OMEN is a Signal Engine, not a Decision Engine.\n    \n    **Migration:**\n    - Use `GET /api/v1/partner-signals/partners` instead\n    - The new endpoint returns the same partner list",
        "operationId": "list_partners_api_v1_partner_risk_partners_get",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Partners Api V1 Partner Risk Partners Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-risk/partners/{symbol}": {
      "get": {
        "tags": [
          "Partner Risk (DEPRECATED)",
          "Partner Risk (DEPRECATED)"
        ],
        "summary": "⚠️ DEPRECATED - Use /api/v1/partner-signals/{symbol}",
        "description": "⚠️ DEPRECATED - Get risk assessment for a specific partner.\n\nThis endpoint violates Signal Engine principles by returning risk verdicts.\n\n**Migration:**\n- Use `GET /api/v1/partner-signals/{symbol}` instead\n- The new endpoint returns signals/metrics\n- RiskCast should make risk decisions",
        "operationId": "get_partner_risk_api_v1_partner_risk_partners__symbol__get",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Partner Risk Api V1 Partner Risk Partners  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/partner-risk/partners/{symbol}/price": {
      "get": {
        "tags": [
          "Partner Risk (DEPRECATED)",
          "Partner Risk (DEPRECATED)"
        ],
        "summary": "⚠️ DEPRECATED - Use /api/v1/partner-signals/{symbol}/price",
        "description": "⚠️ DEPRECATED - Use /api/v1/partner-signals/{symbol}/price instead.",
        "operationId": "get_partner_price_api_v1_partner_risk_partners__symbol__price_get",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Partner Price Api V1 Partner Risk Partners  Symbol  Price Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/partner-risk/webhook/clawbot": {
      "post": {
        "tags": [
          "Partner Risk (DEPRECATED)",
          "Partner Risk (DEPRECATED)"
        ],
        "summary": "⚠️ DEPRECATED - Clawbot integration moved to RiskCast",
        "description": "⚠️ DEPRECATED - Clawbot notification should be triggered by RiskCast, not OMEN.\n\n**Migration:**\n- RiskCast should consume signals from OMEN\n- RiskCast evaluates risk and triggers Clawbot notifications\n- OMEN only provides signals, not notifications",
        "operationId": "trigger_clawbot_notification_api_v1_partner_risk_webhook_clawbot_post",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Clawbot Notification Api V1 Partner Risk Webhook Clawbot Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token (optional, if enabled)"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.omen.io",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.omen.io",
      "description": "Staging server"
...
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}